feat: update structure using domain architecture

+++ lib/core/widgets/get_button.dart
@@ -8,7 +8,7 @@ class GetButton extends StatelessWidget {
   });
 
   final VoidCallback? onPressed;
-  final int? price;
+  final double? price;
 
   @override
   Widget build(BuildContext context) {
